chore(tesseract): split filter logical plan and SQL rendering
Move filter SQL rendering from planner/filter/ to plan/filter/ to align
with the existing planner=logical / plan=physical+SQL boundary. Filter
tree types and analytics stay in planner/filter/tree.rs; ToSql impls
and FilterSqlContext live under plan/filter/.
