+++ rust/cube/cubesqlplanner/cubesqlplanner/src/logical_plan/optimizers/common/helper.rs
@@ -1,7 +1,7 @@
 use itertools::Itertools;
 
 use crate::logical_plan::*;
-use crate::plan::FilterItem;
+use crate::planner::filter::FilterItem;
 use crate::planner::sql_evaluator::MemberSymbol;
 use std::rc::Rc;
