+++ rust/cube/cubesqlplanner/cubesqlplanner/src/logical_plan/pretty_print.rs
@@ -1,4 +1,4 @@
-use crate::plan::FilterItem;
+use crate::planner::filter::FilterItem;
 use crate::planner::sql_evaluator::MemberSymbol;
 use std::rc::Rc;
