+++ rust/cube/cubesqlplanner/cubesqlplanner/src/plan/filter/base_segment.rs
@@ -1,0 +1,22 @@
+use super::ToSql;
+use crate::planner::filter::BaseSegment;
+use crate::planner::query_tools::QueryTools;
+use crate::planner::sql_evaluator::sql_nodes::SqlNode;
+use crate::planner::sql_evaluator::SqlEvaluatorVisitor;
+use crate::planner::sql_templates::PlanSqlTemplates;
+use crate::planner::FiltersContext;
+use cubenativeutils::CubeError;
+use std::rc::Rc;
+
+impl ToSql for BaseSegment {
+    fn to_sql(
+        &self,
+        visitor: &SqlEvaluatorVisitor,
+        node_processor: Rc<dyn SqlNode>,
+        _query_tools: Rc<QueryTools>,
+        templates: &PlanSqlTemplates,
+        _filters_ctx: &FiltersContext,
+    ) -> Result<String, CubeError> {
+        visitor.apply(&self.member_evaluator(), node_processor, templates)
+    }
+}
