+++ rust/cube/cubesqlplanner/cubesqlplanner/src/plan/filter/operators/mod.rs
@@ -2,9 +2,12 @@ pub mod comparison;
 pub mod date_range;
 pub mod date_single;
 pub mod equality;
+pub mod filter_sql_context;
 pub mod in_list;
 pub mod like;
 pub mod measure_filter;
 pub mod nullability;
 pub mod rolling_window;
 pub mod to_date_rolling_window;
+
+pub use filter_sql_context::{FilterOperationSql, FilterSqlContext};
