+++ rust/cube/cubesqlplanner/cubesqlplanner/src/planner/filter/mod.rs
@@ -1,12 +1,15 @@
 pub mod base_filter;
 pub mod base_segment;
 pub mod compiler;
+pub mod filter_debug;
 pub mod filter_operator;
 pub(crate) mod operators;
+pub mod tree;
 pub mod typed_filter;
 
 pub use base_filter::BaseFilter;
 pub use base_segment::BaseSegment;
 pub use filter_operator::FilterOperator;
 pub use operators::date_range::DateRangeOp;
+pub use tree::{Filter, FilterGroup, FilterGroupOperator, FilterItem};
 pub use typed_filter::{resolve_base_symbol, FilterOp};
