+++ rust/cube/cubesqlplanner/cubesqlplanner/src/planner/filter/operators/comparison.rs
@@ -1,6 +1,3 @@
-use super::{FilterOperationSql, FilterSqlContext};
-use cubenativeutils::CubeError;
-
 #[derive(Clone, Debug)]
 pub enum ComparisonKind {
     Gt,
@@ -11,9 +8,9 @@ pub enum ComparisonKind {
 
 #[derive(Clone, Debug)]
 pub struct ComparisonOp {
-    kind: ComparisonKind,
-    value: String,
-    member_type: Option<String>,
+    pub(crate) kind: ComparisonKind,
+    pub(crate) value: String,
+    pub(crate) member_type: Option<String>,
 }
 
 impl ComparisonOp {
@@ -22,18 +19,6 @@ impl ComparisonOp {
             kind,
             value,
             member_type,
-        }
-    }
-}
-
-impl FilterOperationSql for ComparisonOp {
-    fn to_sql(&self, ctx: &FilterSqlContext) -> Result<String, CubeError> {
-        let param = ctx.allocate_and_cast(&self.value, &self.member_type)?;
-        match self.kind {
-            ComparisonKind::Gt => ctx.plan_templates.gt(ctx.member_sql.to_string(), param),
-            ComparisonKind::Gte => ctx.plan_templates.gte(ctx.member_sql.to_string(), param),
-            ComparisonKind::Lt => ctx.plan_templates.lt(ctx.member_sql.to_string(), param),
-            ComparisonKind::Lte => ctx.plan_templates.lte(ctx.member_sql.to_string(), param),
         }
     }
 }
