+++ rust/cube/cubesqlplanner/cubesqlplanner/src/planner/filter/operators/date_single.rs
@@ -1,6 +1,3 @@
-use super::{FilterOperationSql, FilterSqlContext};
-use cubenativeutils::CubeError;
-
 #[derive(Clone, Debug)]
 pub enum DateSingleKind {
     Before,
@@ -11,43 +8,12 @@ pub enum DateSingleKind {
 
 #[derive(Clone, Debug)]
 pub struct DateSingleOp {
-    kind: DateSingleKind,
-    value: String,
+    pub(crate) kind: DateSingleKind,
+    pub(crate) value: String,
 }
 
 impl DateSingleOp {
     pub fn new(kind: DateSingleKind, value: String) -> Self {
         Self { kind, value }
-    }
-}
-
-impl FilterOperationSql for DateSingleOp {
-    fn to_sql(&self, ctx: &FilterSqlContext) -> Result<String, CubeError> {
-        match self.kind {
-            DateSingleKind::Before | DateSingleKind::AfterOrOn => {
-                let param = ctx.format_and_allocate_from_date(&self.value)?;
-                match self.kind {
-                    DateSingleKind::Before => {
-                        ctx.plan_templates.lt(ctx.member_sql.to_string(), param)
-                    }
-                    DateSingleKind::AfterOrOn => {
-                        ctx.plan_templates.gte(ctx.member_sql.to_string(), param)
-                    }
-                    _ => unreachable!(),
-                }
-            }
-            DateSingleKind::BeforeOrOn | DateSingleKind::After => {
-                let param = ctx.format_and_allocate_to_date(&self.value)?;
-                match self.kind {
-                    DateSingleKind::BeforeOrOn => {
-                        ctx.plan_templates.lte(ctx.member_sql.to_string(), param)
-                    }
-                    DateSingleKind::After => {
-                        ctx.plan_templates.gt(ctx.member_sql.to_string(), param)
-                    }
-                    _ => unreachable!(),
-                }
-            }
-        }
     }
 }
