+++ rust/cube/cubesqlplanner/cubesqlplanner/src/planner/filter/operators/equality.rs
@@ -1,11 +1,8 @@
-use super::{FilterOperationSql, FilterSqlContext};
-use cubenativeutils::CubeError;
-
 #[derive(Clone, Debug)]
 pub struct EqualityOp {
-    negated: bool,
-    value: String,
-    member_type: Option<String>,
+    pub(crate) negated: bool,
+    pub(crate) value: String,
+    pub(crate) member_type: Option<String>,
 }
 
 impl EqualityOp {
@@ -14,21 +11,6 @@ impl EqualityOp {
             negated,
             value,
             member_type,
-        }
-    }
-}
-
-impl FilterOperationSql for EqualityOp {
-    fn to_sql(&self, ctx: &FilterSqlContext) -> Result<String, CubeError> {
-        let param = ctx.allocate_and_cast(&self.value, &self.member_type)?;
-        // For negated (notEquals), add OR IS NULL check when value is not null
-        let need_null_check = self.negated;
-        if self.negated {
-            ctx.plan_templates
-                .not_equals(ctx.member_sql.to_string(), param, need_null_check)
-        } else {
-            ctx.plan_templates
-                .equals(ctx.member_sql.to_string(), param, false)
         }
     }
 }
