+++ rust/cube/cubesqlplanner/cubesqlplanner/src/planner/filter/operators/in_list.rs
@@ -1,11 +1,8 @@
-use super::{FilterOperationSql, FilterSqlContext};
-use cubenativeutils::CubeError;
-
 #[derive(Clone, Debug)]
 pub struct InListOp {
-    negated: bool,
-    values: Vec<Option<String>>,
-    member_type: Option<String>,
+    pub(crate) negated: bool,
+    pub(crate) values: Vec<Option<String>>,
+    pub(crate) member_type: Option<String>,
 }
 
 impl InListOp {
@@ -14,22 +11,6 @@ impl InListOp {
             negated,
             values,
             member_type,
-        }
-    }
-}
-
-impl FilterOperationSql for InListOp {
-    fn to_sql(&self, ctx: &FilterSqlContext) -> Result<String, CubeError> {
-        let has_null = self.values.iter().any(|v| v.is_none());
-        let need_null_check = if self.negated { !has_null } else { has_null };
-        let allocated = ctx.allocate_and_cast_values(&self.values, &self.member_type)?;
-
-        if self.negated {
-            ctx.plan_templates
-                .not_in_where(ctx.member_sql.to_string(), allocated, need_null_check)
-        } else {
-            ctx.plan_templates
-                .in_where(ctx.member_sql.to_string(), allocated, need_null_check)
         }
     }
 }
