+++ rust/cube/cubesqlplanner/cubesqlplanner/src/planner/filter/operators/to_date_rolling_window.rs
@@ -1,10 +1,8 @@
-use super::{FilterOperationSql, FilterSqlContext};
 use crate::planner::Granularity;
-use cubenativeutils::CubeError;
 
 #[derive(Clone)]
 pub struct ToDateRollingWindowOp {
-    granularity: Granularity,
+    pub(crate) granularity: Granularity,
 }
 
 impl std::fmt::Debug for ToDateRollingWindowOp {
@@ -18,18 +16,5 @@ impl std::fmt::Debug for ToDateRollingWindowOp {
 impl ToDateRollingWindowOp {
     pub fn new(granularity: Granularity) -> Self {
         Self { granularity }
-    }
-}
-
-impl FilterOperationSql for ToDateRollingWindowOp {
-    fn to_sql(&self, ctx: &FilterSqlContext) -> Result<String, CubeError> {
-        let (from, to) = ctx.date_range_from_time_series()?;
-
-        let from = self
-            .granularity
-            .apply_to_input_sql(ctx.plan_templates, from)?;
-
-        let date_field = ctx.convert_tz(ctx.member_sql)?;
-        ctx.plan_templates.time_range_filter(date_field, from, to)
     }
 }
