+++ rust/cube/cubesqlplanner/cubesqlplanner/src/planner/filter/typed_filter.rs
@@ -1,9 +1,5 @@
-use crate::cube_bridge::member_sql::FilterParamsColumn;
 use crate::planner::query_tools::QueryTools;
-use crate::planner::sql_evaluator::sql_nodes::SqlNode;
-use crate::planner::sql_evaluator::{MemberSymbol, SqlEvaluatorVisitor};
-use crate::planner::sql_templates::PlanSqlTemplates;
-use crate::planner::FiltersContext;
+use crate::planner::sql_evaluator::MemberSymbol;
 use cubenativeutils::CubeError;
 use std::rc::Rc;
 
@@ -18,7 +14,6 @@ use super::operators::measure_filter::MeasureFilterOp;
 use super::operators::nullability::NullabilityOp;
 use super::operators::rolling_window::RegularRollingWindowOp;
 use super::operators::to_date_rolling_window::ToDateRollingWindowOp;
-use super::operators::{FilterOperationSql, FilterSqlContext};
 use super::FilterOperator;
 use crate::planner::GranularityHelper;
 
@@ -97,109 +92,6 @@ impl TypedFilter {
 
     pub fn use_raw_values(&self) -> bool {
         self.use_raw_values
-    }
-
-    pub fn to_sql(
-        &self,
-        visitor: &SqlEvaluatorVisitor,
-        node_processor: Rc<dyn SqlNode>,
-        query_tools: Rc<QueryTools>,
-        plan_templates: &PlanSqlTemplates,
-        filters_ctx: &FiltersContext,
-    ) -> Result<String, CubeError> {
-        if let FilterOp::MeasureFilter(op) = &self.op {
-            return op.to_sql(
-                &self.member_evaluator,
-                visitor,
-                node_processor,
-                query_tools,
-                plan_templates,
-            );
-        }
-
-        let resolved = resolve_base_symbol(&self.member_evaluator);
-        let member_sql = visitor.apply_for_filter(&resolved, node_processor, plan_templates)?;
-
-        let ctx = FilterSqlContext {
-            member_sql: &member_sql,
-            query_tools: &self.query_tools,
-            plan_templates,
-            use_db_time_zone: !filters_ctx.use_local_tz,
-            use_raw_values: self.use_raw_values,
-        };
-
-        self.dispatch_to_sql(&ctx)
-    }
-
-    pub fn to_sql_for_filter_params(
-        &self,
-        column: &FilterParamsColumn,
-        plan_templates: &PlanSqlTemplates,
-        filters_context: &FiltersContext,
-    ) -> Result<String, CubeError> {
-        let use_db_time_zone = !filters_context.use_local_tz;
-
-        match column {
-            FilterParamsColumn::String(column_sql) => {
-                let ctx = FilterSqlContext {
-                    member_sql: column_sql,
-                    query_tools: &self.query_tools,
-                    plan_templates,
-                    use_db_time_zone,
-                    use_raw_values: self.use_raw_values,
-                };
-                self.dispatch_to_sql(&ctx)
-            }
-            FilterParamsColumn::Callback(callback) => {
-                let args = match &self.op {
-                    FilterOp::DateRange(_) | FilterOp::DateSingle(_) => {
-                        let ctx = FilterSqlContext {
-                            member_sql: "",
-                            query_tools: &self.query_tools,
-                            plan_templates,
-                            use_db_time_zone,
-                            use_raw_values: self.use_raw_values,
-                        };
-                        let from = self
-                            .values
-                            .first()
-                            .and_then(|v| v.as_ref())
-                            .map(|v| ctx.format_and_allocate_from_date(v))
-                            .transpose()?;
-                        let to = self
-                            .values
-                            .get(1)
-                            .and_then(|v| v.as_ref())
-                            .map(|v| ctx.format_and_allocate_to_date(v))
-                            .transpose()?;
-                        [from, to].into_iter().flatten().collect()
-                    }
-                    _ => self
-                        .values
-                        .iter()
-                        .filter_map(|v| v.as_ref().map(|v| self.query_tools.allocate_param(v)))
-                        .collect::<Vec<_>>(),
-                };
-                callback.call(&args)
-            }
-        }
-    }
-
-    fn dispatch_to_sql(&self, ctx: &FilterSqlContext) -> Result<String, CubeError> {
-        match &self.op {
-            FilterOp::Comparison(op) => op.to_sql(ctx),
-            FilterOp::DateRange(op) => op.to_sql(ctx),
-            FilterOp::DateSingle(op) => op.to_sql(ctx),
-            FilterOp::Equality(op) => op.to_sql(ctx),
-            FilterOp::InList(op) => op.to_sql(ctx),
-            FilterOp::Like(op) => op.to_sql(ctx),
-            FilterOp::MeasureFilter(_) => {
-                unreachable!("MeasureFilter is handled in TypedFilter::to_sql")
-            }
-            FilterOp::Nullability(op) => op.to_sql(ctx),
-            FilterOp::RegularRollingWindow(op) => op.to_sql(ctx),
-            FilterOp::ToDateRollingWindow(op) => op.to_sql(ctx),
-        }
     }
 }
