+++ rust/cube/cubesqlplanner/cubesqlplanner/src/test_fixtures/test_utils/test_context.rs
@@ -3,7 +3,8 @@ use crate::cube_bridge::join_hints::JoinHintItem;
 use crate::logical_plan::PreAggregationUsage;
 #[cfg(feature = "integration-postgres")]
 use crate::logical_plan::{PreAggregation, PreAggregationSource, PreAggregationTable};
-use crate::plan::Filter;
+use crate::plan::filter::ToSql;
+use crate::planner::filter::Filter;
 use crate::planner::filter::base_segment::BaseSegment;
 use crate::planner::query_tools::QueryTools;
 use crate::planner::sql_evaluator::sql_nodes::SqlNodesFactory;
@@ -649,7 +650,7 @@ impl TestContext {
         let driver_tools = base_tools.driver_tools(false)?;
         let templates = PlanSqlTemplates::try_new(driver_tools, false)?;
 
-        let sql = context.render_filter(&filter, &templates)?;
+        let sql = crate::plan::filter::render_filter(&context, &filter, &templates)?;
         let params = self.query_tools.get_allocated_params();
         Ok((sql, params))
     }
